backend/requirements.txt
fastapi==0.109.2
pydantic==2.6.1
uvicorn==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
pydantic-settings==2.1.0
pydantic[email]>=2.5.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.0.post1
fastapi-cors==0.0.6
colorlog==6.7.0
